LayoutTests/platform/ios-simulator/editing/inserting/insert-3786362-fix-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 5 of DIV > BODY > HTML > #document
EDITING DELEGATE: shouldEndEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 5 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
This demonstrates a bug: copying an empty line that's implemented with a br doesn't put an interchange newline on the pasteboard.  There should be two blank lines after the 'a'.

Before paste:
| "
"
| <div>
|   class="editing"
|   "a"
|   <br>
|   <br>
|   "b"
| "
"
| <div>
|   class="editing"
| "
"

After paste:
| "
"
| <div>
|   class="editing"
|   "a"
|   <br>
|   <br>
|   "b"
| "
"
| <div>
|   class="editing"
| "
"
